+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 9.2.5.35.0 - 2022-xx-xx
+## 9.2.5.35.0 - 2022-06-25
 ### Added
 - Calendar view for earned achievements (only 1st 4 per day for now)
 - Select All and Deselect All buttons for Event Reminders